daemon: constrain max local users to 50
Systems with tens of thousands of users don't want all those users
showing up in the user list.

Set a cap at an even 50, which should cover the lion's share of use
cases well. Of course, if a user not in the list explicitly
logs in (from Not Listed? or whatever) they get added to the list.

https://bugs.freedesktop.org/show_bug.cgi?id=48177