daemon: don't treat explicitly requested users as "cached"
The ListCachedUsers method currently returns users that have
been explicitly requested by a client.  It's weird that merely
querying a user can make it show up in login screen user lists.
Furthermore, UncacheUser is broken since commit
177509e9460b149ecbf85e75c930be2ea00b7d05 because the user has
been explicitly requested in order to uncache it.  So trying
to uncache a user inadvertently caches the user.

This commit fixes that.

https://bugs.freedesktop.org/show_bug.cgi?id=101052