Add support for a list of extra groups for admin users
It's common for distributions to want to add administrator users to
additional groups (e.g. lpadmin or systemd-journal); Debian and Ubuntu
have patches that add this kind of functionality already.

This commit adds a configure option to specify a comma-separated list of
extra groups for admin users and adds support for it when both adding a
new admin user and promoting an user to admin.

https://bugs.freedesktop.org/show_bug.cgi?id=101972