lib: move g_object_ref call for clarity
create_new_user creates an ActUser object and returns it to
the caller.  Before returning the object, it also adds the
object to a list of "new users" to track when it fully loads.

It's idiomatic to return the original ref to the caller of
a constructor, but create_new_user instead gives the original
reference to the new_users list.  Their's no practical difference,
but this commit moves the ref around for clarity.