lib: simplify code dramatically
ActUser is a wrapper over the accountsservice daemon's
managed user objects.  There's a nearly 1-to-1 correspondence
between properties on the proxy to the daemon and properties
on the ActUser object.

This commit dramatically reduces the code, by leveraging the
proxies properties directly, rather than duplicating the values
on the ActUser object.

At the same time, it drops manual GetAll() calls for synchronizing
the proxies properties, since it's completely redundant with the
work the proxy is doing under the hood anyway.

https://bugs.freedesktop.org/show_bug.cgi?id=103488