Expose the automatic login users as a daemon property.
This stops us having to iterate across each user to find which users are
configured for automatic login.

Currently we only support 0 or 1 automatic login user, but this allows
us to support multiple in the future if required. Since D-Bus doesn't have
nullable types [1] this also makes it easier to support the 0 case.

https://bugs.freedesktop.org/show_bug.cgi?id=104564

[1] https://bugs.freedesktop.org/show_bug.cgi?id=27857