lib: don't try to update login frequency manually
The library will try to update the login frequency to 1 if the
requested user isn't finished asynchronously loading yet, but we
know they have an open session.

That no longer works, since we no longer track login-frequency
separately from the dbus proxy object.

This commit drops the code, since it's unnecessary anyway.

To be "on the safe side" we change the value returned for unloaded
users from 0 to 1.  This is okay because the value is undefined
before the user is loaded anyway.