daemon: don't override property from user skeleton
At the moment the skeleton superclass associated with a user keeps
cached copies of all properties, and the derived user class also
keeps duplicate copies of those properties.

The problem is these property values get out of sync and the derived
class fails to emit PropertyChanged signals when the properties
are updated.

This commit solves the issue by cutting out the redundant property
data and just defers all property storage to the skeleton.

https://bugs.freedesktop.org/show_bug.cgi?id=106162