daemon: fix useradd failures caused by invalid read
commit 93e2a85 added support for extra admin groups, but causes useradd
failures, when adding admin accounts. It is because admin_groups are
released before they are actually used, which causes that garbage is
used instead of the group and useradd fails consequently with invalid
group name. Do not release the memory before it is used and also fix
one corresponding memory leak.

https://bugs.freedesktop.org/show_bug.cgi?id=101972