daemon: unregister user after we're done with it not before
Now that we get the object path for a user directly from the skeleton,
we can't access that object path after the skeleton is unregistered.

This commit fixes a problem where the user deletion handling code tries
to access the object path of a user after its skeleton is unregistered.

https://bugs.freedesktop.org/show_bug.cgi?id=106455