daemon: don't emit user-deleted for uncached users
Right now we emit spurious user-deleted signals for
all uncached users anytime a reload occurs.
Uncached users are users explicitly requested by a client,
but not part of the results returned from ListCachedUsers.

We should only be emitting user-deleted it they were initially
cached and transitioned to uncached, not if they're still
hanging around in an uncached state.

This commit fixes the code to do that.

https://bugs.freedesktop.org/show_bug.cgi?id=106455