lib: don't set loaded state until seat is fetched
At the moment we set is-loaded on the user-manager
object as soon as we start fetching the seat, but
we should waiting until the seat is fetched, so
that can_switch() will return the correct value
if the caller waited until the loaded signal
to use it.

This commit changes the >= to > which I believe
was the original intention anyway.

https://bugs.freedesktop.org/show_bug.cgi?id=107298